+++ src/ampio_mqtt/__init__.py
@@ -73,4 +73,4 @@ __all__ = [
     "module_model",
 ]
 
-__version__ = "0.20.0"
+__version__ = "0.21.0"